separate parser from the rest
